requirements.txt
gpustat==0.3.1
pkg-resources==0.0.0
psutil==5.3.0
psycopg2==2.7.3.1
python-decouple==3.1
